Set -fno-PIE by default as a Ubuntu/Mint workaround
Older versions of Ubuntu and its derivates have an issue where the
`file` utility will interpret Position Independent Executables as
shared libraries:

https://bugs.launchpad.net/ubuntu/+source/file/+bug/1747711

As a workaround, allow setting the option COMPILE_WITH_FPIE
(defaults to OFF) and add a suitable message for maintainers in both
the CMake code, during configuration runs and in ccmake.

Update the build instructions in the README to reflect this change.

Closes #94