Bug Fix: Change default build type
In one of the Issues/Discussions it was mentioned that packagers
prefer the Release With Debug builds. Updating the default to use
that so packagers don't have to change anything; this will have
the benefit of giving us that information with our users too.

NOTE: I think it was @ermo that mentioned it.