Peer Review: Travis-CI -> Use Deploy section
Previously marked up the use of the vspackage script as part of
the build section; it really should be in the `before_deploy`
section which only gets called on a deployment.

Section is disabled as it's not ready, so this just stages it for
the next PR to complete it out after the rest of the process is
defined and understood.

Add a packages directory and copy all the packaging files there.