Bug Fix: Git Hub Actions
Split the CI GH Actions from the Release GH Actions to simplify it
a little. This gets rid of some conditional stuff that only applied
to one or the other and allows IS_RELEASE to be hard coded for each
accordingly.

Update the build scripts to use `$var -eq 1` to test IS_RELEASE
as opposed to a dynamic definition.