Bug Fix: CMake version and documentation installation
THe CMake directives originally used for adding the documentation
files to the installation used some parameters only available in
CMake 3.14 and newer. It's a cleaner approach. However, some of the
platforms used still use older version of CMake so the limit is
what is available in CMake 3.5. This makes the adjustment to correct
it so that it still works with CMake 3.5.