Bug Fix: Legacy Data Directory Mode Support
If VS is called with the binary name `vegastrike` then do not
require that a data directory be specified and add the default
locations of the VS:UtCS data directories for use.

If VS is called with the binary name `vegastrike-engine` then
require that the data directory be specified and do not attempt
to search for it.

NOTE: the `vegastrike` command-line should eventually be
transferred to the VS:UtCS assets installation thereby negating
the need for this logic here.