Enhancement: Re-enable dynamic dependency detection
Dynamic dependency detection was fixed and the DEB files built
using `script/package` will properly use it. This re-enables the
CMake/CPack function to rely on it if it exists.

Dynamic dependency detection is preferred as it will make the
dependencies match up with what is actually needed based on an
examination of the binaries even if developers fail to specify it.
This is a more reliable method than static dependencies.