Enhancement: Assets API version
Introduce an Assets API version similar to the Google Android
API Version. It's a simple single value that increments when the
API is changed and can increment indepent of the engine version.

This allows the Assets API to change slightly even between minor
releases, or not change at all across major releases, and to
communicate that via a simple mechanism.