Initial commit - refactor cargo code
- Add cargo_hold class
- Merge product into cargo class. no reason not to.
- Delete store. Not used.

A cargo hold is a single container of cargo items in the ship. It can be of type cargo, upgrade or hidden for smuggling.

With this change, we mostly decouple cargo code from the rest of the game. We move mass and money handling (pointer) to ComponentsManager, which is really a proto ship.